Introduce AbstractNotificationDriver and move getPublicDriverId into it.
